requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dash==3.0.4
dash-bootstrap-components==2.0.2
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.39.0
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
plotly==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
